requirements.txt
dealer==2.0.5
django-bootstrap3==11.0.0
django-debug-toolbar==1.10.1
django-js-reverse==0.8.2
django-mptt==0.9.1
django-widget-tweaks==1.4.3
Django==2.2.9
djangorestframework-camel-case==0.2.0
djangorestframework==3.8.2
mysqlclient
six==1.11.0
Pygments==2.2.0
Pillow
chardet==3.0.4
mime==0.1.0
django-otp==0.6.0
django-two-factor-auth==1.10.0
lark-parser==0.6.6
shortuuid==0.5.0
